% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{list_colorblind_safe}
\alias{list_colorblind_safe}
\title{List all colorblind-safe palettes}
\usage{
list_colorblind_safe()
}
\value{
A data frame with columns \code{journal}, \code{palette}, \code{n_colors}.
}
\description{
List all colorblind-safe palettes
}
\examples{
list_colorblind_safe()
}
