% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{list_journals}
\alias{list_journals}
\title{List all available journal keys}
\usage{
list_journals()
}
\value{
A character vector of journal keys, sorted alphabetically.
}
\description{
List all available journal keys
}
\examples{
list_journals()
}
