% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{pubchroma-package}
\alias{pubchroma}
\alias{pubchroma-package}
\title{pubchroma: Journal-Inspired Color Palettes for Scientific Figures}
\description{
Provides color palettes inspired by major scientific journals (Nature, Science, Cell, NEJM, Lancet, JAMA, PNAS, BMJ) for use in biomedical and engineering figures. Includes colorblind-safety checking and integration with ggplot2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tyuan2024/pubchroma}
  \item Report bugs at \url{https://github.com/tyuan2024/pubchroma/issues}
}

}
\author{
\strong{Maintainer}: tyuan2024 \email{tyuan2024@126.com}

}
\keyword{internal}
